Remove cyclic Rust dev-dependencies
